apps/atlas/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "kbve_atlas" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/atlas/html' --cov-report xml:'../../coverage/apps/atlas/coverage.xml' --html='../../reports/apps/atlas/unittests/html/index.html' --junitxml='../../reports/apps/atlas/unittests/junit.xml'"

[tool.poetry]
name = "kbve"
version = "1.0.2"
description = "ATLAS"
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "kbve_atlas"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.11"
  pyautogen = {extras = ["blendsearch"], version = "^0.1.6"}
  openai = "^0.28.1"
  litellm = "^0.1.824"
  docker = "^6.1.3"
  beautifulsoup4 = "^4.12.2"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
